packages/api/pyproject.toml
[project]
name = "techconnect-api"
version = "0.1.0"
description = "FastAPI API for TechConnect biomedical research application"
requires-python = ">=3.14.2"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.128.2",
    "uvicorn[standard]>=0.40.0",
    "techconnect-schemas",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "httpx>=0.28.1",
    "ruff>=0.15.0",
    "pyrefly>=0.51.0",
]

[project.scripts]
seed-db = "app.seed:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.pyrefly]
project-includes = ["."]
project-excludes = ["**/.[!/.]*", "**/__pycache__"]
